+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to the "cli-wrapper" will be documented in this file.
 - Update `package.json` in publishable projects to include packaging scripts and other smaller changes.
 - Update tsconfig lib to use `es2023` instead of `es2022`.
 - Remove references to `@gmjs/tsconfig`, `@gmjs/prettier-config` and `gmjs/jest-config` from generated code and the project, and use direct configurations instead.
+- Use `vitest` instead of `jest` in react generated projects.
 - Other generated files improvements and fixes.
 
 ### Added